fix: random and just build
