+++ .gitignore
@@ -19,3 +19,5 @@ node_modules
 **/.svelte-kit/*
 
 .DS_Store
+
+.artifact/
